+++ PhotoWork/src/gAPainter/Shape.java
@@ -5,6 +5,6 @@ import java.awt.Graphics;
 public interface Shape {
 
 	public void mutate();
-	public void paint(Graphics g, int scale);
+	public void paint(Graphics g, double scale);
 	public void paint(Graphics g);
 }
